+++ sphinx/ext/autodoc/directive.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Callable
+from collections.abc import Callable
+from typing import TYPE_CHECKING, Any
 
 from docutils import nodes
 from docutils.statemachine import StringList